requirements.txt
aniso8601==9.0.1
astroid==2.11.5
autopep8==1.6.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Login==0.6.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==1.0.1
gcloud==0.18.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
mccabe==0.7.0
oauth2client==4.1.3
packaging==21.3
platformdirs==2.5.2
protobuf==4.21.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pylint==2.13.9
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.20.0
python-jwt==3.3.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-toolbelt==0.9.1
rsa==4.8
safrs==2.12.2
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
